switch to BigDecimal

+++ app/src/main/java/com/example/acalc/CalcFunctions.java
@@ -1,6 +1,8 @@
 package com.example.acalc;
 
+import java.math.BigDecimal;
+
 public enum CalcFunctions {
-    SINE {public double evaluate(double operand){ return Math.sin(operand);}};
-    public abstract double evaluate(double operand);
+    SINE {public BigDecimal evaluate(BigDecimal operand){ return BigDecimal.valueOf(Math.sin(operand.doubleValue()));}};
+    public abstract BigDecimal evaluate(BigDecimal operand);
 }
